Finished Version
This is the finished version of my project. I have a completed menu, upgrade system, boss fight, win and death menus, option menus, enemy level up systems and a fading system for when you load into the game. I also have a completed pause menu and options menu with info on the controls and a button to toggle the music.